+++ libs/angular-swa-auth/src/index.ts
@@ -9,3 +9,4 @@ export * from './lib/auth-event-persistence.service';
 export * from './lib/identity-provider-selector.service';
 export * from './lib/first-identity-provider-selector.service';
 export * from './lib/identity-provider-interactive-selector.service';
+export * from './lib/swa-roles-check.directive';