+++ libs/angular-swa-auth/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - `SwaRoleGuard`: add to a route to perform an authorization check using roles defined in Status Web App
+- `SwaRoleCheckDirective`: Structural directive to perform an authorization check using roles defined in Status Web App 
 
 ## [1.0.0] - 2021-08-26